fix: 修复所有 clippy 警告
- 修复文档注释问题（empty_line_after_doc_comments）
- 优化代码风格（redundant_field_names, collapsible_if, redundant_closure 等）
- 性能优化（clone_on_copy, ptr_arg, unnecessary_lazy_evaluations 等）
- 代码改进（manual_range_contains, unnecessary_map_or, manual_strip 等）
- 特殊处理（too_many_arguments, io_other_error 等）

共修复 67 个 clippy 警告，提升代码质量和可读性

+++ src/id_generator.rs
@@ -23,7 +23,7 @@ impl IdGenerator {
     /// 从字节数组创建 ID 生成器
     pub fn from_bytes(machine_id_bytes: [u8; 6]) -> Self {
         let raw_worker_id = u16::from_be_bytes([machine_id_bytes[0], machine_id_bytes[1]]);
-        let worker_id = (raw_worker_id % 1024) as u16;
+        let worker_id = raw_worker_id % 1024;
         Self {
             worker_id,
             sequence: AtomicU64::new(0),
